refactor(cli): harden parse_int, add doc comments, fix const-cast
- parse_int: narrow params from long to int (callers already pass
  int-range literals; eliminates truncation on cast to *out)
- engine.h: document generate.stream ownership (caller retains)
- cmds.c sigint_handler: document async-signal-safety invariant
- cmds.c slurp_stdin: document strip-all-trailing-newlines intent
- test_cli_cmds: replace const-cast fmemopen with char[] (finding #7)

+++ src/cli/cli.c
@@ -16,7 +16,7 @@ static cli_cmd_t cli_error(cli_args_t *out, const char *msg) {
 }
 
 /* Parse a whole-string integer in [min, max]. Returns 0 on success. */
-static int parse_int(const char *s, long min, long max, int *out) {
+static int parse_int(const char *s, int min, int max, int *out) {
     char *end = NULL;
     errno = 0;
     long v = strtol(s, &end, 10);
